refactor constants for clarity
